refactor: rename validator files to align with CLI subcommands
- validator.py → validate.py, validator_report.py → validate_render.py
- CLI: flatten to `validate` and `validate-render` top-level commands
- Improve docs examples with practical usage patterns
